backend/requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
requests>=2.31.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
